requirements.txt

langchain
langchain-community
langchain-ollama
langchain-groq
gradio
python-dotenv
tavily-python
wikipedia
langchain-tavily
pytest